Add the Image classes and drawImage() methods to Canvas. (#691)
* Add the Image and TextureProxy classes.

* Add the Size struct.

* Add MakeWH(int, int) to Rect.

* Fix auto tests.

* Add drawImage() to Canvas.

* Add assignCacheOwner() to Resource.

* Fix build error on web platform.

+++ src/base/PreComposeLayer.cpp
@@ -91,7 +91,6 @@ bool PreComposeLayer::verify() const {
 }
 
 Rect PreComposeLayer::getBounds() const {
-  return Rect::MakeWH(static_cast<float>(composition->width),
-                      static_cast<float>(composition->height));
+  return Rect::MakeWH(composition->width, composition->height);
 }
 }  // namespace pag
